Add Serpent depletion conversion script
Creates an equivalent depletion chain using decay and fission yield
files distributed with Serpent. Some steps are taken to clean up and
fix a few issues. The fixes are detailed in the script and here for
completeness.

First, the files appear to be ENDF-6 formatted files concatenated
together. Unfortunately, there is a TEND record after each dataset,
so :func:`openmc.data.endf.get_evaluations` will only read and return
the first evaluation. For this reason, we must read the file until
all records have been found manually.

Secondly, the decay file does not have a tape indent (TPID) record,
which can throw off the first reading. This exception can be caught,
but it currently involves skipping the first item in the file. At the
time of this writing, this skipped item contains the decay data for a
neutron.

Lastly, there is a single negative uncertainty in the Co-55 k-shell
conversion energies. This quantity is not used in the depletion chain,
but it must be altered to prevent downstream failures.

Supports reading in an additional chain file and applying branching
ratios prior to exiting.